Fix to unconflate index of reuse vs index of next available child.
A link reused to link to a parent must be handled distinctly, which requires it to be at a fixed
index, versus handling of any next child, versus handling of further children.

Previously, I'd used the term "first" and conflated it to mean both fixed-index handling for reuse
and any-next handling.  That doesn't work, when the first of any next child can be at a variable
index.