Change bound on Link type to new Link trait (instead of BorrowMut).
This avoids imposing a requirement of using some outside trait.  This now requires that an
appropriate implementation of our own trait must be provided by the user, which avoids having to
abuse, or mistakenly reuse, an impl of a std trait.

BorrowMut was often already impl'ed for common link and node types, for both owned values and
references, which was a little more convenient for those cases.  But it's possible that a
preexisting impl is inappropriate for the needs of deep_safe_drop, or that an impl made for those
needs is inappropriate for all of the intent of BorrowMut.

It's better to not assume that any premade trait is always appropriate.  It's easy for every use
of deep_safe_drop to impl our new Link trait, and this forces the user to think more about how
exactly that should be.